requirements.txt
Django==4.0.4
djangorestframework==3.13.1
mysqlclient
gunicorn
redis
django-redis
